requirements.txt
Flask==2.3.3 # Or your preferred Flask version
mysql-connector-python==8.0.33 # Or a compatible version
python-dotenv==1.0.0 # Or a compatible version